scheda tomba e scheda individui update to fix bug

+++ modules/db/structures/SCHEDAIND_table.py
@@ -4,7 +4,7 @@ Created on 19 feb 2018
 @author: Serena Sensini
 '''
 
-from sqlalchemy import Table, Column, Integer, Float, String, Text, MetaData, create_engine, UniqueConstraint
+from sqlalchemy import Table, Column, Integer, Float, Numeric,String, Text, MetaData, create_engine, UniqueConstraint
 
 from modules.db.pyarchinit_conn_strings import Connection
 
@@ -37,13 +37,13 @@ class SCHEDAIND_table:
                             Column('completo_si_no', String(5)),
                             Column('disturbato_si_no', String(5)),
                             Column('in_connessione_si_no', String(5)),
-                            Column('lunghezza_scheletro', Float(2, 2)),
+                            Column('lunghezza_scheletro', Numeric(2,2)),
                             Column('posizione_scheletro', String(50)),
                             Column('posizione_cranio', String(50)),
                             Column('posizione_arti_superiori', String(50)),
                             Column('posizione_arti_inferiori', String(50)),
                             Column('orientamento_asse', Text),
-                            Column('orientamento_azimut', Float(2, 2)),
+                            Column('orientamento_azimut', Numeric(2,2)),
 
                             # explicit/composite unique constraint.  'name' is optional.
                             UniqueConstraint('sito', 'nr_individuo', name='ID_individuo_unico')
